+++ XenAdmin/Controls/DataGridViewEx/CollapsingPoolHostRowSorter.cs
@@ -1,0 +1,75 @@
+﻿/* Copyright (c) Citrix Systems, Inc. 
+ * All rights reserved. 
+ * 
+ * Redistribution and use in source and binary forms, 
+ * with or without modification, are permitted provided 
+ * that the following conditions are met: 
+ * 
+ * *   Redistributions of source code must retain the above 
+ *     copyright notice, this list of conditions and the 
+ *     following disclaimer. 
+ * *   Redistributions in binary form must reproduce the above 
+ *     copyright notice, this list of conditions and the 
+ *     following disclaimer in the documentation and/or other 
+ *     materials provided with the distribution. 
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND 
+ * CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, 
+ * INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF 
+ * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE 
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR 
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, 
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, 
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR 
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS 
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING 
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE 
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF 
+ * SUCH DAMAGE.
+ */
+
+using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using XenAdmin.Core;
+
+namespace XenAdmin.Controls.DataGridViewEx
+{
+    /// <summary>
+    /// A class implementation to do sorting of CollapsingPoolHostDataGridViewRow derived types based on the input column index and direction
+    public class CollapsingPoolHostRowSorter<T> : CollapsingPoolHostDataGridViewRowStableSorter<T> where T : CollapsingPoolHostDataGridViewRow
+    {
+        private int columnClicked;
+
+        public CollapsingPoolHostRowSorter(ListSortDirection direction, int columnClicked)
+            : base(direction)
+        {
+            this.columnClicked = columnClicked;
+
+            // In case of same string, default sorting should be still in order under Descending direction
+            StableSorter = new CollapsingPoolHostDataGridViewRowDefaultSorter(Direction);
+        }
+
+        protected override int SortRowByColumnDetails(T leftSide, T rightSide)
+        {
+            if (leftSide.IsPoolOrStandaloneHost && !rightSide.IsPoolOrStandaloneHost)
+                return -1;
+
+            if (!leftSide.IsPoolOrStandaloneHost && rightSide.IsPoolOrStandaloneHost)
+                return 1;
+
+            if ((leftSide.IsPoolOrStandaloneHost && rightSide.IsPoolOrStandaloneHost) ||
+                (!leftSide.IsPoolOrStandaloneHost && !rightSide.IsPoolOrStandaloneHost))
+            {
+                return StringUtility.NaturalCompare(leftSide.Cells[columnClicked].Value.ToString(),
+                    rightSide.Cells[columnClicked].Value.ToString());
+            }
+
+            return 0;
+        }
+    }
+}
